# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Next Clear Look Engine API
  version: 1.0.0
  summary: Sentinel-2 geometric opportunities, AOI archive evidence and historical likelihoods.
  description: |
    HTTP contract between the engine and the web app. An opportunity is geometry, not promised acquisition.
    A likelihood is empirical history, not a weather forecast. All instants are aware UTC RFC 3339.
servers:
  - url: http://localhost:8000/v1
    description: Local engine
...
paths:
...
  /aois:
    get:
      tags: [AOIs]
      operationId: listAois
      summary: List saved AOIs.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: AOI page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AoiPage'
              examples:
                saved:
                  externalValue: ./examples/aoi-page.json
    post:
      tags: [AOIs]
      operationId: createAoi
      summary: Validate, canonicalise and save a Polygon or MultiPolygon.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AoiCreate'
            examples:
              singapore:
                externalValue: ./examples/aoi-create.json
      responses:
        '201':
          description: AOI created.
          headers:
            Location:
              schema: {type: string, format: uri-reference}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aoi'
              examples:
                singapore:
                  externalValue: ./examples/aoi.json
        '409':
          description: Idempotency key was reused with different content.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /aois/{aoi_id}:
    parameters:
      - $ref: '#/components/parameters/AoiId'
    get:
      tags: [AOIs]
      operationId: getAoi
      summary: Get one AOI.
      responses:
        '200':
          description: AOI.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aoi'
              examples:
                singapore:
                  externalValue: ./examples/aoi.json
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [AOIs]
      operationId: updateAoi
      summary: Rename or replace AOI geometry; geometry changes invalidate derived cache keys.
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/AoiPatch'
            examples:
              rename:
                externalValue: ./examples/aoi-patch.json
      responses:
        '200':
          description: Updated AOI.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aoi'
              examples:
                singapore:
                  externalValue: ./examples/aoi.json
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [AOIs]
      operationId: deleteAoi
      summary: Delete an AOI and its unpinned derived records.
      responses:
        '204':
          description: Deleted; no response body.
        '404':
          $ref: '#/components/responses/NotFound'
...
  /analysis-jobs:
    get:
      tags: [Jobs]
      operationId: listAnalysisJobs
      summary: List jobs, newest first.
      parameters:
        - name: state
          in: query
          schema:
            $ref: '#/components/schemas/JobState'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Job page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobPage'
              examples:
                recent:
                  externalValue: ./examples/analysis-jobs.json
    post:
      tags: [Jobs]
      operationId: createAnalysisJob
      summary: Queue bounded, cancellable analysis work.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisJobCreate'
            examples:
              full_analysis:
                externalValue: ./examples/analysis-job-create.json
      responses:
        '202':
          description: Accepted, or an identical existing job was returned.
          headers:
            Location:
              schema: {type: string, format: uri-reference}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
              examples:
                singapore:
                  externalValue: ./examples/analysis-job.json
        '409':
          description: Idempotency conflict or incompatible mode.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /analysis-jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Jobs]
      operationId: getAnalysisJob
      summary: Get durable job state and terminal result/error.
      responses:
        '200':
          description: Job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
              examples:
                completed:
                  externalValue: ./examples/analysis-job-completed.json
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Jobs]
      operationId: cancelAnalysisJob
      summary: Idempotently request cancellation.
      responses:
        '202':
          description: Cancellation requested or job already cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
              examples:
                cancelling:
                  externalValue: ./examples/analysis-job-cancelling.json
        '409':
          description: Job already reached a non-cancelled terminal state.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /analysis-jobs/{job_id}/events:
    parameters:
      - $ref: '#/components/parameters/JobId'
      - $ref: '#/components/parameters/LastEventIdQuery'
    get:
      tags: [Streams]
      operationId: streamAnalysisJobEvents
      summary: Resumeable per-job progress, warning, partial-result and terminal events.
      parameters:
        - $ref: '#/components/parameters/LastEventIdHeader'
      responses:
        '200':
          description: SSE stream. Schemas and ordering are normative in events.md.
          headers:
            Cache-Control:
              schema: {type: string, const: no-cache}
            X-Accel-Buffering:
              schema: {type: string, const: 'no'}
          content:
            text/event-stream:
              schema: {type: string}
              examples:
                replay:
                  externalValue: ./examples/job-events.sse
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Requested event is older than retained job history.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    AoiId:
      name: aoi_id
      in: path
      required: true
      schema: {type: string, pattern: '^aoi_[A-Za-z0-9_-]+$'}
...
    JobId:
      name: job_id
      in: path
      required: true
      schema: {type: string, pattern: '^job_[A-Za-z0-9_-]+$'}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    Cursor:
      name: cursor
      in: query
      schema: {type: string, maxLength: 512}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 8, maxLength: 128}
    LastEventIdHeader:
      name: Last-Event-ID
      in: header
      required: false
      schema: {type: string, maxLength: 200}
    LastEventIdQuery:
      name: last_event_id
      in: query
      required: false
      schema: {type: string, maxLength: 200}
      description: Browser-compatible fallback when a client cannot set Last-Event-ID; header wins.
...
  responses:
    NotFound:
      description: The addressed resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing:
              externalValue: ./examples/error.json
    ValidationError:
      description: Request syntax was valid but one or more values failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: The bounded job queue is full or a source rate limit applies.
      headers:
        Retry-After:
          schema: {type: integer, minimum: 1}
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Timestamp:
      type: string
      format: date-time
      examples: ['2026-10-02T17:43:54Z']
    GeoJsonGeometry:
      oneOf:
        - $ref: '#/components/schemas/GeoJsonPolygon'
        - $ref: '#/components/schemas/GeoJsonMultiPolygon'
    Position:
      type: array
      prefixItems:
        - {type: number, minimum: -180, maximum: 180}
        - {type: number, minimum: -90, maximum: 90}
      minItems: 2
      maxItems: 2
    LinearRing:
      type: array
      items: {$ref: '#/components/schemas/Position'}
      minItems: 4
    GeoJsonPolygon:
      type: object
      additionalProperties: false
      required: [type, coordinates]
      properties:
        type: {type: string, const: Polygon}
        coordinates:
          type: array
          items: {$ref: '#/components/schemas/LinearRing'}
          minItems: 1
    GeoJsonMultiPolygon:
      type: object
      additionalProperties: false
      required: [type, coordinates]
      properties:
        type: {type: string, const: MultiPolygon}
        coordinates:
          type: array
          minItems: 1
          items:
            type: array
            minItems: 1
            items: {$ref: '#/components/schemas/LinearRing'}
    Bbox:
      type: array
      prefixItems:
        - {type: number}
        - {type: number}
        - {type: number}
        - {type: number}
      minItems: 4
      maxItems: 4
    GeoJsonPoint:
      type: object
      additionalProperties: false
      required: [type, coordinates]
      properties:
        type: {type: string, const: Point}
        coordinates: {$ref: '#/components/schemas/Position'}
    AoiPreset:
      type: object
      additionalProperties: false
      required: [slug, locality, story, climate_tags]
      properties:
        slug: {type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'}
        locality: {type: string}
        story: {type: string}
        climate_tags:
          type: array
          items: {type: string}
    ReplayCoverage:
      type: object
      additionalProperties: false
      required: [opportunities, archive, likelihood, thumbnails]
      properties:
        opportunities: {type: boolean}
        archive: {type: boolean}
        likelihood: {type: boolean}
        thumbnails: {type: boolean}
    AoiCreate:
      type: object
      additionalProperties: false
      required: [name, geometry]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        geometry: {$ref: '#/components/schemas/GeoJsonGeometry'}
        timezone: {type: string, default: UTC, maxLength: 80}
    AoiPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        geometry: {$ref: '#/components/schemas/GeoJsonGeometry'}
        timezone: {type: string, maxLength: 80}
    Aoi:
      type: object
      additionalProperties: false
      required: [id, name, origin, preset, geometry, centroid, bbox, timezone, area_km2, geometry_sha256, replay_coverage, created_at, updated_at, provenance_id]
      properties:
        id: {type: string}
        name: {type: string}
        origin: {type: string, enum: [preset, user]}
        preset:
          oneOf: [{$ref: '#/components/schemas/AoiPreset'}, {type: 'null'}]
        geometry: {$ref: '#/components/schemas/GeoJsonGeometry'}
        centroid: {$ref: '#/components/schemas/GeoJsonPoint'}
        bbox: {$ref: '#/components/schemas/Bbox'}
        timezone: {type: string}
        area_km2: {type: number, exclusiveMinimum: 0, maximum: 250000}
        geometry_sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        replay_coverage: {$ref: '#/components/schemas/ReplayCoverage'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
        provenance_id: {type: string}
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [count, next_cursor]
      properties:
        count: {type: integer, minimum: 0}
        next_cursor: {type: [string, 'null']}
        as_of: {$ref: '#/components/schemas/Timestamp'}
    AoiPage:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Aoi'}}
        meta: {$ref: '#/components/schemas/PaginationMeta'}
...
    JobState:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    JobType:
      type: string
      enum: [full_analysis, orbit_only, archive_refresh, raster_scene, likelihood_only]
    JobStage:
      type: string
      enum: [catalogue, orbit, archive, raster, likelihood, finalise]
    AnalysisJobCreate:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type: {$ref: '#/components/schemas/JobType'}
        aoi_id: {type: string}
        scene_id: {type: string}
        from: {$ref: '#/components/schemas/Timestamp'}
        to: {$ref: '#/components/schemas/Timestamp'}
        force_refresh: {type: boolean, default: false}
      anyOf:
        - required: [aoi_id]
        - required: [scene_id]
    JobResult:
      type: object
      additionalProperties: true
      properties:
        opportunity_count: {type: integer, minimum: 0}
        scene_count: {type: integer, minimum: 0}
        likelihood_url: {type: string, format: uri-reference}
    AnalysisJob:
      type: object
      additionalProperties: false
      required: [id, type, state, progress, stage, created_at, result, error, events_url]
      properties:
        id: {type: string}
        type: {$ref: '#/components/schemas/JobType'}
        aoi_id: {type: [string, 'null']}
        scene_id: {type: [string, 'null']}
        state: {$ref: '#/components/schemas/JobState'}
        progress: {type: number, minimum: 0, maximum: 1}
        stage: {$ref: '#/components/schemas/JobStage'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        started_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        finished_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        cancel_requested_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        result:
          oneOf: [{$ref: '#/components/schemas/JobResult'}, {type: 'null'}]
        error:
          oneOf: [{$ref: '#/components/schemas/ErrorObject'}, {type: 'null'}]
        events_url: {type: string, format: uri-reference}
    JobPage:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/AnalysisJob'}}
        meta: {$ref: '#/components/schemas/PaginationMeta'}
...
    ErrorObject:
      type: object
      additionalProperties: false
      required: [code, message, retryable, details, request_id]
      properties:
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]+$'}
        message: {type: string}
        retryable: {type: boolean}
        details: {type: object, additionalProperties: true}
        request_id: {type: string}
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorObject'}